move Spec conversion to value_codec.cpp
